% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqTab_categorical.R
\name{freqTab_categorical}
\alias{freqTab_categorical}
\title{Frequency Table for Categorical variables}
\usage{
freqTab_categorical(
  datasetColumn,
  numberCategories,
  classColname = "class",
  cutDig.lab = 0
)
}
\arguments{
\item{datasetColumn}{Designed for nominal and ordinal categorical variables from a cleansed dataset column, such as: cars$speed}

\item{numberCategories}{The number of categories that handles the variable + 1. E.g.: 4 Categories; Fully agree (...) Fully disagree; numberCategories = 5.}

\item{classColname}{Optional: Provide a string of text to rename (1, 1) indexes from the table generated. Default name: "class".}

\item{cutDig.lab}{Provides an integer to dig.lab parameter from cut() function in case classes are rational.}
}
\value{
A frequency table to interpret or to calculate measures of central tendency.
}
\description{
Made for nominal and ordinal categorical variables, generates a frequency table which contains the following calculations: Frequency, Relative Frequency, Accumulated Frequency and Relative Accumulated Frequency.
}
\examples{
diamonds <- ggplot2::diamonds
cut <- freqTab_categorical(diamonds$cut, 6, "cut")
cut
}
