% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqTab_elements.R
\name{freqTab_elements}
\alias{freqTab_elements}
\title{Elements composing a Frequency Table}
\usage{
freqTab_elements(datasetColumn)
}
\arguments{
\item{datasetColumn}{From a cleansed dataset column, such as: diamonds$price}
}
\value{
A list with number of observations, classes according to Sturges formula, amplitude of every class and limits of every group
}
\description{
Generates a list with the elements composing your Frequency table generated by the freqTab_quantitative() and freqTab_categorical() functions. It is also used to calculate measures of central tendency.
WARNING: It is important to note, that data columns that has 0 as its min() will not function.
}
\examples{
diamonds <- ggplot2::diamonds
freqTab_elements(diamonds$price)
}
