% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqTab_quantitative.R
\name{freqTab_quantitative}
\alias{freqTab_quantitative}
\title{Frequency Table for quantitative variables}
\usage{
freqTab_quantitative(datasetColumn, classColname = "class", cutDig.lab = 0)
}
\arguments{
\item{datasetColumn}{From a cleansed dataset column, such as: diamonds$price}

\item{classColname}{Optional: Provide a string of text to rename (1, 1) indexes from the table generated. Default name: "class".}

\item{cutDig.lab}{Provides an integer to dig.lab parameter from cut() function in case classes are rational.}
}
\value{
A frequency table to interpret or to calculate measures of central tendency.
}
\description{
Made for quantitative variables, generates a frequency table which contains the following calculations: Frequency, Relative Frequency, Accumulated Frequency and Relative Accumulated Frequency.
WARNING: It is important to note, that data columns that has 0 as its min() will not function.
}
\examples{
diamonds <- ggplot2::diamonds
price <- freqTab_quantitative(diamonds$price, "price", 10)
price
}
