\name{freqtable}
\alias{freqtable}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Creates nice frequency tables for Data prasentation
}
\description{
Create frequency tables for Output files and Markdown files
}
\usage{
freqtable(Var, F1)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Var}{
Variable
}
  \item{foot}{
character vector
}
}
\author{
Lukas Kraiger
}
\examples{
#freqtable()
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
