% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_freqTab.R
\name{median_freqTab}
\alias{median_freqTab}
\title{Median for grouped data}
\usage{
median_freqTab(
  datasetColumn,
  freqTable.Freq,
  freqTable.Acc.Freq,
  round = T,
  digits
)
}
\arguments{
\item{datasetColumn}{From the original cleansed dataset column, such as: cars$speed}

\item{freqTable.Freq}{Frequency column from the table generated by freqTab functions. E.g.: diamonds$Freq}

\item{freqTable.Acc.Freq}{Accumulated Frequency column from the table generated by freqTab functions. E.g.: diamonds$Acc.Freq}

\item{round}{Boolean operator that indicates whether the Median is rounded or not. Default is TRUE.}

\item{digits}{Digits for the Median  to be rounded with the round() function. Default at the round() default being 0 for a whole number.}
}
\value{
Median from the grouped data.
}
\description{
A function which calculates the Median for grouped data. It uses the table generated by freqTab_categorical() or freqTab_quantitative().
}
\examples{
diamonds <- ggplot2::diamonds
price <- freqTab_quantitative(diamonds$price, "price", 10)
median_price <- median_freqTab(diamonds$price, price$Freq, price$Acc.Freq)
median_price
}
