% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_freqTab.R
\name{mode_freqTab}
\alias{mode_freqTab}
\title{Mode for grouped data}
\usage{
mode_freqTab(datasetColumn, freqTable.Freq, round = T, digits)
}
\arguments{
\item{datasetColumn}{From a cleansed dataset column, such as: diamonds$price}

\item{freqTable.Freq}{Frequency column from the table generated by freqTab functions. E.g.: diamonds$Freq}

\item{round}{Boolean operator that indicates whether the Median is rounded or not. Default is TRUE.}

\item{digits}{Digits for the Median  to be rounded with the round() function. Default at the round() default being 0 for a whole number.}
}
\value{
Returns the Mode as integer.
}
\description{
A function which calculates the Mode for grouped data. It uses the table generated by freqTab_categorical() or freqTab_quantitative().
}
\examples{
diamonds <- ggplot2::diamonds
price <- freqTab_quantitative(diamonds$price, "price", 10)
mode_price <- mode_freqTab(diamonds$price, price$Freq)
mode_price
}
